# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/users:
    get:
      operationId: listChatUsers
      tags: [chat]
      summary: Search app users to start a chat with
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChatUsersPayload"
...
  /chat/conversations/support:
    post:
      operationId: startSupportConversation
      tags: [chat]
      summary: Start or fetch the current user's support conversation
      responses:
        "200":
          description: Existing conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSupportConversationPayload"
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSupportConversationPayload"
  /chat/conversations/direct:
    post:
      operationId: startDirectConversation
      tags: [chat]
      summary: Start or fetch a direct conversation with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDirectConversationInput"
      responses:
        "200":
          description: Existing conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartDirectConversationPayload"
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartDirectConversationPayload"
  /chat/conversations/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChatMessagesPayload"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendChatMessagePayload"
  /chat/conversations/{id}/read:
    patch:
      operationId: markConversationRead
      tags: [chat]
      summary: Mark a conversation as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Marked read
  /chat/messages/{id}:
    delete:
      operationId: deleteChatMessage
      tags: [chat]
      summary: Delete a chat message (for me or for everyone)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [me, everyone]
            default: me
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
  /chat/messages/{id}/forward:
    post:
      operationId: forwardChatMessage
      tags: [chat]
      summary: Forward a message to another conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForwardChatMessageInput"
      responses:
        "201":
          description: Forwarded message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendChatMessagePayload"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Hadith interactions ────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorPayload:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ChatUser:
      type: object
      required: [id, clerkUserId, name, email, updatedAt, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: string }
        avatarUrl: { type: ["string", "null"] }
        updatedAt: { type: string }
        createdAt: { type: string }
    ListChatUsersPayload:
      type: array
      items:
        $ref: "#/components/schemas/ChatUser"
...
    StartSupportConversationPayload:
      type: object
      required: [id]
      properties:
        id: { type: integer }
    StartDirectConversationInput:
      type: object
      required: [targetClerkId]
      properties:
        targetClerkId: { type: string }
    StartDirectConversationPayload:
      type: object
      required: [id]
      properties:
        id: { type: integer }
    ChatMessage:
      type: object
      required: [id, conversationId, senderClerkId, senderName, senderIsAdmin, isDeleted, isForwarded, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        senderClerkId: { type: string }
        senderName: { type: string }
        senderNickname: { type: ["string", "null"] }
        senderAvatarUrl: { type: ["string", "null"] }
        senderIsAdmin:
          type: string
          enum: ["true", "false"]
        content: { type: ["string", "null"] }
        attachmentUrl: { type: ["string", "null"] }
        attachmentType:
          type: ["string", "null"]
          enum: [image, video, file, null]
        attachmentName: { type: ["string", "null"] }
        attachmentMimeType: { type: ["string", "null"] }
        attachmentSize: { type: ["integer", "null"] }
        isDeleted: { type: boolean }
        isForwarded: { type: boolean }
        forwardedFromSenderName: { type: ["string", "null"] }
        createdAt: { type: string }
    ListChatMessagesPayload:
      type: array
      items:
        $ref: "#/components/schemas/ChatMessage"
    ChatMessageInput:
      type: object
      properties:
        content: { type: string }
        attachmentUrl: { type: string }
        attachmentType:
          type: string
          enum: [image, video, file]
        attachmentName: { type: string }
        attachmentMimeType: { type: string }
        attachmentSize: { type: integer }
    SendChatMessagePayload:
      $ref: "#/components/schemas/ChatMessage"
    ForwardChatMessageInput:
      type: object
      required: [targetConversationId]
      properties:
        targetConversationId: { type: integer }

    # ── Storage ────────────────────────────────────────────────────────
...